Fix large contract deployments signature

+++ app/rust/include/rslib.h
@@ -11,7 +11,7 @@ uint16_t rs_c32_address(const uint8_t *input, uint8_t version, uint8_t *output, 
 
 /****************************** others ***********************************************************/
 
-parser_error_t _parser_init(parser_context_t *ctx, const uint8_t *buffer, uint16_t bufferSize, uint16_t *alloc_size);
+parser_error_t _parser_init(parser_context_t *ctx, const uint8_t *buffer, uint32_t bufferSize, uint16_t *alloc_size);
 
 parser_error_t _read(const parser_context_t *c, parser_tx_t *v);
